add menagement view

+++ src/main/java/ShopAppBackend/FrontMainPageManagement/Section/SectionRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SectionRepository extends JpaRepository<Section,Long> {
+public interface SectionRepository extends JpaRepository<Section,Integer> {
 }
